Update list of available babel plugins.
When you use a spread inside of a destructuring, babel will not correctly
compile that and leaves the spread in the destructuring. Since that's not
compatible yet, we must include transform-es2015-destructuring.

If you use a spread inside of a destructuring inside of a parameter list,
babel will also not correctly compile that. At first we though we could
add back in the trasnform-es2015-parameters plugin, and for a while that
fixed it. Hovwer, the transform-es2015-parameters plugin combined with the
transform-class-properties plugin will create an incorrect reference to
`this` in all bound instance method defined with the es7 syntax. Since
this is a far more common use case, the transform-es2015-parameters plugin
has been turned off